syncing changes

+++ ICE-3/ExpressJSTest/controllers/studentsController.js
@@ -5,11 +5,11 @@ const Student = require("../models/studentModel"); //mongoose student structure 
 const mongoose = require('mongoose');
 
 
-// const { get } = require("../routes/studentRoute");
+//const { get } = require("../routes/studentRoute");
 
 //@access public
 const getStudents = asyncHandler(async (req, res) => {
-    const students = await Student.find();
+    const students = await Student.find({ user_id: req.user.id });
     res.status(200.).json(students);
 });
 
@@ -24,16 +24,30 @@ const createStudent = asyncHandler(async (req, res) => {
         name,
         email,
         phone,
+        user_id: req.user.id,
     });
     res.status(201).json(student);
 });
 
 
+//@desc Update a studnets
+//@route PUT /api/students/:id
+//@access private
 const updateStudent = asyncHandler(async (req, res) => {
-    const student = await Student.findById(req.params.id);
+    const studentId = req.params.id;
+    const isValidId = mongoose.Types.ObjectId.isValid(studentId);
+    if (!isValidId) {
+        res.status(400);
+        throw new Error("Id is not valid");
+    }
+    const student = await Student.findById(studentId);
     if (!student) {
         res.status(404);
         throw new Error("Contact Not Found");
+    }
+    if (student.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("User do not have permission to update other user's students!");
     }
     const updatedStudent = await Student.findByIdAndUpdate(
         req.params.id,
@@ -42,14 +56,26 @@ const updateStudent = asyncHandler(async (req, res) => {
     );
     res.status(200.).json(updatedStudent);
 });
-
+//@desc Delete a student
+//@route DELETE /api/students/:id
+//@access private
 const deleteStudent = asyncHandler(async (req, res) => {
-    const student = await Student.findById(req.params.id);
+    const studentId = req.params.id;
+    const isValidId = mongoose.Types.ObjectId.isValid(studentId);
+    if (!isValidId) {
+        res.status(400);
+        throw new Error("Id is not valid");
+    }
+    const student = await Student.findById(studentId);
     if (!student) {
         res.status(404);
-        throw new Error("Contact Not Found");
+        throw new Error("Student Not Found");
     }
-    await Student.deleteOne({ _id: req.params.id }); // changed to deleteOne(); remove(); was not an available function
+    if (student.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("User do not have permission to delete other user's students!");
+    }
+    await Student.deleteOne({ _id: req.params.id });
     res.status(200.).json(student);
 });
